+++ ggml/src/ggml-sycl/set_rows.cpp
@@ -63,12 +63,12 @@ static void set_rows_sycl_q(const char * __restrict__ src0_d,
         const int64_t i12         = i03 % ne12;
         const int64_t i11         = i02 % ne11;
         const int64_t i10         = i01;
-        const size_t  src_offset  = calculate_offset<3>({ nb01, nb02, nb03 }, { i01, i02, i03 });
+        const size_t  src_offset  = calculate_offset<3>({ (int64_t)nb01, (int64_t)nb02, (int64_t)nb03 }, { i01, i02, i03 });
         const char *  src_block   = src0_d + src_offset + i00 * sizeof(float);
-        const size_t  src1_offset = calculate_offset<3>({ nb10, nb11, nb12 }, { i10, i11, i12 });
+        const size_t  src1_offset = calculate_offset<3>({ (int64_t)nb10, (int64_t)nb11, (int64_t)nb12 }, { i10, i11, i12 });
         const int64_t dst_row     = src1_d[src1_offset / sizeof(TIdx)];
         const size_t  dst_offset =
-            calculate_offset<3>({ nb1, nb2, nb3 }, { dst_row, i02, i03 }) + (i00 / qk) * sizeof(blockType);
+            calculate_offset<3>({ (int64_t)nb1, (int64_t)nb2, (int64_t)nb3 }, { dst_row, i02, i03 }) + (i00 / qk) * sizeof(blockType);
         char * dst_block = reinterpret_cast<char *>(reinterpret_cast<char *>(dst_d) + dst_offset);
         cpyblck(src_block, dst_block);
     });
@@ -104,9 +104,9 @@ static void k_set_rows(
     const int64_t i11 = i02 % ne11;
     const int64_t i10 = i01;
 
-    const int64_t dst_row = *(const TIdx *)((const char *)src1 + calculate_offset<3>({nb10, nb11, nb12}, {i10, i11, i12}));
+    const int64_t dst_row = *(const TIdx *)((const char *)src1 + calculate_offset<3>({(int64_t)nb10, (int64_t)nb11, (int64_t)nb12}, {i10, i11, i12}));
 
-    const char * src0_row = src0 + calculate_offset<3>({nb01, nb02, nb03}, {i01, i02, i03});
+    const char * src0_row = src0 + calculate_offset<3>({(int64_t)nb01, (int64_t)nb02, (int64_t)nb03}, {i01, i02, i03});
     const char * src_elem = src0_row + i00 * src_type_size;
     char * dst_row_ptr = dst + dst_row*nb1 + i02*nb2 + i03*nb3;
     char * dst_elem = dst_row_ptr + i00 * dst_type_size;
